refactor(pusher): split runtime ownership seams
Verification: 98 focused Pusher/finalization tests passed; backend typecheck passed; Black and diff checks passed.
