+++ backend/routers/pusher.py
@@ -1,29 +1,37 @@
 import struct
 import asyncio
-import json
 import time
 from collections import deque
-from typing import Any, Awaitable, Deque, Dict, List, Optional, TypedDict, TypeVar, cast
+from typing import Any, Awaitable, Dict, List, Optional, cast
 
 from fastapi import APIRouter
 from fastapi.websockets import WebSocketDisconnect, WebSocket
 from starlette.websockets import WebSocketState
 
 import database.conversations as conversations_db
-from database import conversation_finalization_jobs as finalization_jobs_db
 from database import users as users_db
-from services.conversation_finalization import final_attempt_failed
+from routers.pusher_finalization import process_conversation_task
+from routers.pusher_protocol import (
+    BUFFERED_AUDIO_MAX_BYTES,
+    MAX_SAMPLE_RATE,
+    MIN_SAMPLE_RATE,
+    PRIVATE_CLOUD_QUEUE_MAX_SIZE,
+    AudioBytesQueueItem,
+    ByteBudget,
+    PrivateCloudChunk,
+    SpeakerSampleRequest,
+    TranscriptQueueItem,
+    append_bounded,
+    bound_private_pending,
+    extend_bounded,
+    frame_header,
+    json_object,
+    pusher_session_outcome,
+)
 from utils.apps import is_audio_bytes_app_enabled
 from utils.app_integrations import (
     trigger_realtime_integrations,
     trigger_realtime_audio_bytes,
-)
-from utils.byok import set_byok_keys, set_byok_uid
-from utils.conversations import lifecycle as lifecycle_service
-from utils.conversations.finalizer import (
-    ConversationFinalizationDisposition,
-    ConversationFinalizationError,
-    finalize_persisted_conversation,
 )
 from utils.executors import db_executor, storage_executor, run_blocking, start_background_task
 from utils.async_tasks import (
@@ -37,16 +45,14 @@ from utils.webhooks import (
     realtime_transcript_webhook,
     get_audio_bytes_webhook_seconds,
 )
-from utils.cloud_tasks import get_listen_finalization_tasks_max_attempts, is_audio_merge_dispatch_enabled
+from utils.cloud_tasks import is_audio_merge_dispatch_enabled
 from utils.other.storage import maybe_invalidate_conversation_playback, upload_audio_chunks_batch
 from utils.metrics import (
     PUSHER_ACTIVE_WS_CONNECTIONS,
     PUSHER_PRIVATE_CLOUD_UPLOAD_DROPS,
-    PUSHER_QUEUE_DROPPED_BYTES,
-    PUSHER_QUEUE_DROPS,
 )
 from utils.readiness import ReadinessGate
-from utils.observability.journeys import JourneyAttempt, JourneyOutcome, record_capture_finalization_terminal
+from utils.observability.journeys import JourneyAttempt
 from utils.speaker_identification import extract_speaker_samples
 import logging
 
@@ -65,7 +71,6 @@ PRIVATE_CLOUD_BATCH_MAX_AGE = 60.0  # seconds — flush batch if oldest chunk ex
 PRIVATE_CLOUD_SYNC_MAX_RETRIES = 3
 
 # Queue size limits
-PRIVATE_CLOUD_QUEUE_MAX_SIZE = 20  # ~18MB/connection max (30 conns × 18MB = 540MB) — prevents OOM with headroom
 SPEAKER_SAMPLE_QUEUE_WARN_SIZE = 100
 
 # Constants for transcript queue batching
@@ -85,121 +90,6 @@ WS_RECEIVE_TIMEOUT = 300.0  # seconds
 # before being force-cancelled.  Prevents hung GCS uploads or webhook calls from
 # blocking cleanup indefinitely.
 BG_DRAIN_TIMEOUT = 30.0  # seconds
-MIN_SAMPLE_RATE = 8000
-MAX_SAMPLE_RATE = 48000
-BUFFERED_AUDIO_MAX_BYTES = 20 * 1024 * 1024
-PRIVATE_CLOUD_PENDING_MAX_CONVERSATIONS = PRIVATE_CLOUD_QUEUE_MAX_SIZE
-
-_QueueItem = TypeVar('_QueueItem')
-
-
-class _ByteBudget:
-    def __init__(self, limit: int):
-        self.limit = limit
-        self.used = 0
-
-    def reserve(self, size: int) -> bool:
-        if size < 0 or self.used + size > self.limit:
-            return False
-        self.used += size
-        return True
-
-    def release(self, size: int) -> None:
-        self.used = max(0, self.used - size)
-
-
-def _record_queue_drop(queue_name: str, size: int = 0) -> None:
-    PUSHER_QUEUE_DROPS.labels(queue=queue_name).inc()
-    if size:
-        PUSHER_QUEUE_DROPPED_BYTES.labels(queue=queue_name).inc(size)
-
-
-def _append_bounded(
-    queue: Deque[_QueueItem],
-    item: _QueueItem,
-    queue_name: str,
-    *,
-    byte_budget: Optional[_ByteBudget] = None,
-    size_of: Optional[Any] = None,
-) -> bool:
-    dropped = len(queue) == queue.maxlen
-    if dropped:
-        removed = queue.popleft()
-        removed_size = size_of(removed) if size_of else 0
-        if byte_budget:
-            byte_budget.release(removed_size)
-        _record_queue_drop(queue_name, removed_size)
-    queue.append(item)
-    return dropped
-
-
-def _extend_bounded(buffer: bytearray, data: bytes, queue_name: str, byte_budget: _ByteBudget) -> bool:
-    if not byte_budget.reserve(len(data)):
-        _record_queue_drop(queue_name, len(data))
-        return False
-    buffer.extend(data)
-    return True
-
-
-def _bound_private_pending(pending: Dict[str, Dict[str, Any]], byte_budget: _ByteBudget) -> None:
-    if len(pending) < PRIVATE_CLOUD_PENDING_MAX_CONVERSATIONS:
-        return
-    oldest_id = next(iter(pending))
-    removed = pending.pop(oldest_id)
-    removed_size = len(removed['data'])
-    byte_budget.release(removed_size)
-    _record_queue_drop('private_cloud_pending', removed_size)
-
-
-def _frame_header(data: bytes) -> int:
-    if len(data) < 4:
-        raise ValueError('frame header is incomplete')
-    header_type = struct.unpack('<I', data[:4])[0]
-    if header_type not in {100, 101, 102, 103, 104, 105}:
-        raise ValueError('unknown frame type')
-    if header_type == 101 and len(data) < 12:
-        raise ValueError('audio frame is incomplete')
-    return header_type
-
-
-def _json_object(data: bytes) -> Dict[str, Any]:
-    value = json.loads(bytes(data[4:]).decode("utf-8"))
-    if not isinstance(value, dict):
-        raise ValueError('frame payload must be an object')
-    return value
-
-
-def pusher_session_outcome(close_code: int, *, application_failed: bool = False) -> JourneyOutcome:
-    """Classify accepted sessions without counting normal disconnects as failures."""
-    if application_failed or close_code == 1011:
-        return 'failure'
-    if close_code in {1000, 1001}:
-        return 'success'
-    return 'cancelled'
-
-
-class _SpeakerSampleRequest(TypedDict):
-    person_id: str
-    conversation_id: str
-    segment_ids: List[str]
-    queued_at: float
-
-
-class _TranscriptQueueItem(TypedDict):
-    segments: List[Dict[str, Any]]
-    memory_id: Optional[str]
-
-
-class _AudioBytesQueueItem(TypedDict):
-    type: str
-    sample_rate: int
-    data: bytearray
-
-
-class _PrivateCloudChunk(TypedDict):
-    data: bytes
-    conversation_id: str
-    timestamp: float
 
 
 async def _dispatch_transcript_item(uid: str, segments: List[Dict[str, Any]], memory_id: Optional[str]) -> None:
@@ -213,199 +103,6 @@ async def _dispatch_transcript_item(uid: str, segments: List[Dict[str, Any]], me
         run('integrations', trigger_realtime_integrations(uid, segments, memory_id)),
         run('webhook', realtime_transcript_webhook(uid, segments)),
     )
-
-
-async def _process_conversation_task(
-    uid: str,
-    conversation_id: str,
-    language: str,
-    websocket: WebSocket,
-    byok_keys: Optional[Dict[str, str]] = None,
-    finalization_job_id: Optional[str] = None,
-    dispatch_generation: Optional[int] = None,
-) -> None:
-    """Process a leased conversation job and send a minimal result to listen.
-
-    `byok_keys` is forwarded from the listen service. When present, LLM and
-    STT calls made inside process_conversation route through the user's own
-    provider keys instead of Omi's env keys.
-    """
-    if byok_keys:
-        set_byok_keys(byok_keys)
-        set_byok_uid(uid)
-
-    async def send_result(result: Dict[str, Any]) -> None:
-        """Attempt the optional live acknowledgement after durable work.
-
-        The Firestore finalization transition is authoritative. A listener can
-        close after handing opcode 104 to pusher, so a failed result write must
-        never turn an already-completed durable job into a worker failure.
-        """
-        data = bytearray()
-        data.extend(struct.pack("I", 201))
-        data.extend(bytes(json.dumps(result), "utf-8"))
-        try:
-            await websocket.send_bytes(bytes(data))
-        except (RuntimeError, WebSocketDisconnect):
-            logger.info(
-                'pusher finalization result undeliverable after source close uid=%s conversation=%s',
-                uid,
-                conversation_id,
-            )
-
-    job_id: Optional[str] = None
-    generation: Optional[int] = None
-    lease_epoch: Optional[int] = None
-    attempt_count: int = 0
-
-    async def record_failure(failure_code: str) -> bool:
-        """Release the lease. Returns whether this was the terminal attempt.
-
-        Inline dispatch has no Cloud Tasks worker to exhaust the attempt budget,
-        so the claimed attempt count is the only bound on a deterministically
-        failing job. Without a terminal state the conversation would stay
-        `processing` forever and be re-finalized by every later session.
-        """
-        if job_id is None or generation is None or lease_epoch is None:
-            return False
-        terminal = attempt_count >= get_listen_finalization_tasks_max_attempts()
-        try:
-            if terminal:
-                marked_dead_letter = await run_blocking(
-                    db_executor,
-                    final_attempt_failed,
-                    job_id,
-                    generation,
-                    lease_epoch,
-                    attempt_count,
-                )
-                if not marked_dead_letter:
-                    return False
-                return True
-            await run_blocking(
-                db_executor,
-                finalization_jobs_db.mark_finalization_retryable,
-                job_id,
-                generation,
-                lease_epoch,
-                failure_code,
-            )
-        except Exception:
-            logger.error(
-                'pusher finalization recovery update failed uid=%s conversation=%s failure=%s terminal=%s',
-                uid,
-                conversation_id,
-                failure_code,
-                terminal,
-            )
-            return False
-        return False
-
-    try:
-        if not finalization_job_id or dispatch_generation is None:
-            # Every finalization request must be mediated by the Firestore
-            # owner.  Accepting the legacy frame would allow a pending pusher
-            # session to bypass the durable claim and double-process work.
-            await send_result({'conversation_id': conversation_id, 'error': 'durable_job_required'})
-            return
-
-        job_id = finalization_job_id
-        generation = dispatch_generation
-
-        claim = await run_blocking(
-            db_executor,
-            finalization_jobs_db.claim_finalization_job,
-            job_id,
-            generation,
-            allow_byok=bool(byok_keys),
-            expected_uid=uid,
-            expected_conversation_id=conversation_id,
-        )
-        claim_status = claim['status']
-        if claim_status == 'fenced':
-            await send_result({'conversation_id': conversation_id, 'fenced': True})
-            return
-        if claim_status == 'completed':
-            await send_result({'conversation_id': conversation_id, 'success': True})
-            return
-        if claim_status != 'claimed':
-            await send_result(
-                {
-                    'conversation_id': conversation_id,
-                    'error': f'job_{claim_status}',
-                    # A dead-lettered job is never actionable again; telling the
-                    # live session it is terminal stops it from re-requesting.
-                    'terminal': claim_status in finalization_jobs_db.TERMINAL_JOB_STATUSES,
-                }
-            )
-            return
-        attempt_count = claim['attempt_count']
-        lease_epoch = claim['lease_epoch']
-        if lease_epoch is None:
-            logger.error(
-                'pusher finalization claim returned no lease epoch uid=%s conversation=%s', uid, conversation_id
-            )
-            await send_result({'conversation_id': conversation_id, 'error': 'processing_failed'})
-            return
-
-        disposition = await finalize_persisted_conversation(
-            uid,
-            conversation_id,
-            language,
-            finalization_job_id=job_id,
-            dispatch_generation=generation,
-            lease_epoch=lease_epoch,
-        )
-
-        if disposition == ConversationFinalizationDisposition.fenced:
-            completed = await run_blocking(
-                db_executor,
-                lifecycle_service.complete_fenced_finalization,
-                job_id,
-                generation,
-                lease_epoch,
-            )
-        else:
-            completed = await run_blocking(
-                db_executor,
-                finalization_jobs_db.mark_finalization_completed,
-                job_id,
-                generation,
-                lease_epoch,
-            )
-        if not completed:
-            await send_result({'conversation_id': conversation_id, 'error': 'job_completion_conflict'})
-            return
-        if disposition == ConversationFinalizationDisposition.fenced:
-            record_capture_finalization_terminal('stale', claim.get('created_at'))
-            await send_result({'conversation_id': conversation_id, 'fenced': True})
-            return
-        record_capture_finalization_terminal('success', claim.get('created_at'))
-        await send_result({'conversation_id': conversation_id, 'success': True})
-    except ConversationFinalizationError:
-        terminal = await record_failure('processing_failed')
-        logger.error(
-            'pusher finalization failed uid=%s conversation=%s failure=processing_failed terminal=%s',
-            uid,
-            conversation_id,
-            terminal,
-        )
-        try:
-            await send_result({'conversation_id': conversation_id, 'error': 'processing_failed', 'terminal': terminal})
-        except Exception:
-            pass
-    except Exception:
-        terminal = await record_failure('worker_failed')
-        logger.error(
-            'pusher finalization task failed uid=%s conversation=%s failure=worker_failed terminal=%s',
-            uid,
-            conversation_id,
-            terminal,
-        )
-        try:
-            await send_result({'conversation_id': conversation_id, 'error': 'processing_failed', 'terminal': terminal})
-        except Exception:
-            pass
 
 
 async def _websocket_util_trigger(
@@ -463,16 +160,16 @@ async def _websocket_util_trigger(
         raise
 
     # Bounded queues — prevent unbounded memory growth during backpressure
-    speaker_sample_queue: deque[_SpeakerSampleRequest] = deque(maxlen=SPEAKER_SAMPLE_QUEUE_WARN_SIZE)
-    transcript_queue: deque[_TranscriptQueueItem] = deque(maxlen=TRANSCRIPT_QUEUE_WARN_SIZE)
-    audio_bytes_queue: deque[_AudioBytesQueueItem] = deque(maxlen=AUDIO_BYTES_QUEUE_WARN_SIZE)
+    speaker_sample_queue: deque[SpeakerSampleRequest] = deque(maxlen=SPEAKER_SAMPLE_QUEUE_WARN_SIZE)
+    transcript_queue: deque[TranscriptQueueItem] = deque(maxlen=TRANSCRIPT_QUEUE_WARN_SIZE)
+    audio_bytes_queue: deque[AudioBytesQueueItem] = deque(maxlen=AUDIO_BYTES_QUEUE_WARN_SIZE)
 
     # private_cloud_queue caps at PRIVATE_CLOUD_QUEUE_MAX_SIZE to prevent OOM kills.
     # An OOM kill loses ALL queued data for ALL users on the pod — dropping the oldest
     # chunk for one user is strictly better than killing the pod.
-    private_cloud_queue: deque[_PrivateCloudChunk] = deque(maxlen=PRIVATE_CLOUD_QUEUE_MAX_SIZE)
+    private_cloud_queue: deque[PrivateCloudChunk] = deque(maxlen=PRIVATE_CLOUD_QUEUE_MAX_SIZE)
     audio_bytes_event = asyncio.Event()  # Signals when items are added for instant wake
-    audio_budget = _ByteBudget(BUFFERED_AUDIO_MAX_BYTES)
+    audio_budget = ByteBudget(BUFFERED_AUDIO_MAX_BYTES)
 
     async def process_private_cloud_queue() -> None:
         """Background task that batches private cloud sync uploads by conversation_id.
@@ -488,10 +185,10 @@ async def _websocket_util_trigger(
         # Pending batches keyed by conversation_id
         pending: Dict[str, Dict[str, Any]] = {}
 
-        def _add_to_batch(chunk_info: _PrivateCloudChunk) -> None:
+        def _add_to_batch(chunk_info: PrivateCloudChunk) -> None:
             conv_id = chunk_info['conversation_id']
             if conv_id not in pending:
-                _bound_private_pending(pending, audio_budget)
+                bound_private_pending(pending, audio_budget)
                 pending[conv_id] = {
                     'data': bytearray(),
                     'conversation_id': conv_id,
@@ -616,8 +313,8 @@ async def _websocket_util_trigger(
             # Separate ready and pending requests.
             # On shutdown, skip the age check — process everything so pending
             # samples aren't silently dropped when the drain timeout fires.
-            ready_requests: List[_SpeakerSampleRequest] = []
-            pending_requests: List[_SpeakerSampleRequest] = []
+            ready_requests: List[SpeakerSampleRequest] = []
+            pending_requests: List[SpeakerSampleRequest] = []
 
             for request in list(speaker_sample_queue):
                 if is_shutdown or current_time - request['queued_at'] >= SPEAKER_SAMPLE_MIN_AGE:
@@ -657,7 +354,7 @@ async def _websocket_util_trigger(
                 continue
 
             # Process batch
-            batch: List[_TranscriptQueueItem] = list(transcript_queue)
+            batch: List[TranscriptQueueItem] = list(transcript_queue)
             transcript_queue.clear()
 
             for item in batch:
@@ -682,7 +379,7 @@ async def _websocket_util_trigger(
                 continue
 
             # Process all queued items
-            batch: List[_AudioBytesQueueItem] = list(audio_bytes_queue)
+            batch: List[AudioBytesQueueItem] = list(audio_bytes_queue)
             audio_bytes_queue.clear()
 
             for item in batch:
@@ -721,7 +418,7 @@ async def _websocket_util_trigger(
                     websocket_close_code = 1011
                     application_failed = True
                     break
-                header_type = _frame_header(data)
+                header_type = frame_header(data)
 
                 # Heartbeat (data-frame keepalive from backend to reset GKE ILB idle timer)
                 if header_type == 100:
@@ -742,7 +439,7 @@ async def _websocket_util_trigger(
                                 f"private_cloud_queue full ({len(private_cloud_queue)}/{PRIVATE_CLOUD_QUEUE_MAX_SIZE}), "
                                 f"dropping oldest chunk to prevent OOM {uid}"
                             )
-                        _append_bounded(
+                        append_bounded(
                             private_cloud_queue,
                             {
                                 'data': bytes(private_cloud_sync_buffer),
@@ -764,7 +461,7 @@ async def _websocket_util_trigger(
 
                 # Transcript - queue for batched processing
                 if header_type == 102:
-                    res = _json_object(data)
+                    res = json_object(data)
                     segments = res.get('segments')
                     memory_id = res.get('memory_id')
                     if not isinstance(segments, list) or not all(isinstance(segment, dict) for segment in segments):
@@ -780,7 +477,7 @@ async def _websocket_util_trigger(
                     # Route this transcript by its own memory_id when present, falling back to the
                     # session's conversation id. This does not mutate session-scoped state.
                     conversation_or_memory_id = memory_id or current_conversation_id
-                    _append_bounded(
+                    append_bounded(
                         transcript_queue,
                         {'segments': segments, 'memory_id': conversation_or_memory_id},
                         'transcript',
@@ -789,7 +486,7 @@ async def _websocket_util_trigger(
 
                 # Process conversation request
                 if header_type == 104:
-                    res = _json_object(data)
+                    res = json_object(data)
                     conversation_id = res.get('conversation_id')
                     language = res.get('language', 'en')
                     byok_keys = res.get('byok_keys') or None
@@ -817,7 +514,7 @@ async def _websocket_util_trigger(
                         # this handoff, so pusher owns it at process scope. Its
                         # lifespan drains tracked tasks on process shutdown.
                         start_background_task(
-                            _process_conversation_task(
+                            process_conversation_task(
                                 uid,
                                 conversation_id,
                                 language,
@@ -832,7 +529,7 @@ async def _websocket_util_trigger(
 
                 # Speaker sample extraction request - queue for background processing
                 if header_type == 105:
-                    res = _json_object(data)
+                    res = json_object(data)
                     person_id = res.get('person_id')
                     conv_id = res.get('conversation_id')
                     segment_ids = res.get('segment_ids', [])
@@ -850,7 +547,7 @@ async def _websocket_util_trigger(
                         logger.info(
                             f"Queued speaker sample request: person={person_id}, {len(segment_ids)} segments {uid}"
                         )
-                        _append_bounded(
+                        append_bounded(
                             speaker_sample_queue,
                             {
                                 'person_id': person_id,
@@ -871,9 +568,9 @@ async def _websocket_util_trigger(
                     # Only accumulate audio buffers if there's a consumer (app trigger or webhook)
                     # Without this guard, buffers grow ~16KB/s indefinitely for users with no audio apps
                     if has_audio_apps_enabled:
-                        _extend_bounded(trigger_audiobuffer, audio_data, 'audio_app_buffer', audio_budget)
+                        extend_bounded(trigger_audiobuffer, audio_data, 'audio_app_buffer', audio_budget)
                     if audio_bytes_webhook_delay_seconds is not None:
-                        _extend_bounded(audiobuffer, audio_data, 'audio_webhook_buffer', audio_budget)
+                        extend_bounded(audiobuffer, audio_data, 'audio_webhook_buffer', audio_budget)
 
                     # Private cloud sync - queue chunks for background processing
                     if private_cloud_sync_enabled and current_conversation_id:
@@ -881,7 +578,7 @@ async def _websocket_util_trigger(
                             # Use timestamp from first buffer of this 5-second chunk
                             private_cloud_chunk_start_time = buffer_start_timestamp
 
-                        _extend_bounded(
+                        extend_bounded(
                             private_cloud_sync_buffer,
                             audio_data,
                             'private_cloud_buffer',
@@ -894,7 +591,7 @@ async def _websocket_util_trigger(
                                     f"private_cloud_queue full ({len(private_cloud_queue)}/{PRIVATE_CLOUD_QUEUE_MAX_SIZE}), "
                                     f"dropping oldest chunk to prevent OOM {uid}"
                                 )
-                            _append_bounded(
+                            append_bounded(
                                 private_cloud_queue,
                                 {
                                     'data': bytes(private_cloud_sync_buffer),
@@ -915,7 +612,7 @@ async def _websocket_util_trigger(
                     ):
                         if len(audio_bytes_queue) >= AUDIO_BYTES_QUEUE_WARN_SIZE:
                             logger.warning(f"Warning: audio_bytes_queue size {len(audio_bytes_queue)} {uid}")
-                        _append_bounded(
+                        append_bounded(
                             audio_bytes_queue,
                             {
                                 'type': 'app',
@@ -934,7 +631,7 @@ async def _websocket_util_trigger(
                     ):
                         if len(audio_bytes_queue) >= AUDIO_BYTES_QUEUE_WARN_SIZE:
                             logger.warning(f"Warning: audio_bytes_queue size {len(audio_bytes_queue)} {uid}")
-                        _append_bounded(
+                        append_bounded(
                             audio_bytes_queue,
                             {
                                 'type': 'webhook',
@@ -949,7 +646,7 @@ async def _websocket_util_trigger(
                         audiobuffer = bytearray()
                     continue
 
-        except (ValueError, struct.error, UnicodeDecodeError, json.JSONDecodeError) as exc:
+        except (ValueError, struct.error, UnicodeDecodeError) as exc:
             websocket_close_code = 1003
             logger.warning(f'Invalid pusher frame: {type(exc).__name__} {uid}')
         except WebSocketDisconnect as exc:
@@ -967,7 +664,7 @@ async def _websocket_util_trigger(
                         f"private_cloud_queue full ({len(private_cloud_queue)}/{PRIVATE_CLOUD_QUEUE_MAX_SIZE}), "
                         f"dropping oldest chunk to prevent OOM {uid}"
                     )
-                _append_bounded(
+                append_bounded(
                     private_cloud_queue,
                     {
                         'data': bytes(private_cloud_sync_buffer),
